Fix bug, change to correct string
