+++ src/main/java/ar/com/api/general/dto/SimplePriceFilterDTO.java
@@ -11,7 +11,6 @@ import java.util.Optional;
 @Builder
 public class SimplePriceFilterDTO implements IFilter, Serializable {
 
-
     @NotBlank
     @NotEmpty
     private String ids;
@@ -40,7 +39,7 @@ public class SimplePriceFilterDTO implements IFilter, Serializable {
                     .append(include24hrVol.isPresent() ? include24hrVol.get() : "false")
                 .append("&include_24hr_change=")
                     .append(include24hrChange.isPresent() ? include24hrChange.get() : "false")
-                .append("&=include_last_updated_at")
+                .append("&include_last_updated_at=")
                     .append(includeLastUpdatedAt.isPresent() ? includeLastUpdatedAt.get() : "false")
                 .append("&precision=")
                     .append(precision.isPresent() ? precision.get() : "0");
